% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_distance.R
\name{manhattan_dist}
\alias{manhattan_dist}
\title{This is a function that calculates the Manhattan distance between
two points in a plane.
It takes four inputs: the x and y coordinates of Point 1 and the x and y coordinates of Point 2.}
\usage{
manhattan_dist(x_1, y_1, x_2, y_2)
}
\arguments{
\item{x_1}{number}

\item{y_1}{number}

\item{x_2}{number}

\item{y_2}{number}
}
\value{
The Manhattan distance of the points x: \code{x_1}, \code{x_2} and points y: \code{y_1}, \code{y_2}
}
\description{
This is a function that calculates the Manhattan distance between
two points in a plane.
It takes four inputs: the x and y coordinates of Point 1 and the x and y coordinates of Point 2.
}
