% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean_plot.R
\name{plot_euclidean_dist}
\alias{plot_euclidean_dist}
\title{This function plots the Euclidean distance between 2 points.
It takes four inputs: the x and y coordinates of Point 1 and the x and y coordinates of Point 2.}
\usage{
plot_euclidean_dist(x_1, x_2, y_1, y_2)
}
\arguments{
\item{x_1}{number}

\item{x_2}{number}

\item{y_1}{number}

\item{y_2}{number}
}
\value{
A single line plot showing Euclidean distance of the points x: \code{x_1}, \code{x_2} and points y: \code{y_1}, \code{y_2}
}
\description{
This function plots the Euclidean distance between 2 points.
It takes four inputs: the x and y coordinates of Point 1 and the x and y coordinates of Point 2.
}
